SILGen: do not emit hop_to_executor in some objc thunks
An @objc async thunk does not need to perform a
hop_to_executor if the native method the thunk is
wrapping is async, since the native method will do
its own hop.

This redundant hop was causing a crash in SILGen
with async actor-instance isolated @objc methods
declared in an actor, because SILGen was not
prepared to add the hop to  to `self` within
the @objc thunk.

resolves rdar://80130628