[Runtime] Static constructors and destructors should be an error here.
Except for a handful of places we know about, there should be no static
constructors or destructors in the runtime; they're undesirable because
they inflate start up or shut down times, and in the case of destructors
we can't even guarantee that they will actually run (e.g. if the program
declares that it supports fast termination, they just won't).

It should be OK to change the warnings to errors as a result.

rdar://80965245