[stdlib] Set a more correct deployment target for stdlib binaries
On Apple platforms, parts of the standard library are tightly coupled to specific versions of libobjc/Foundation, so the stdlib binaries built from this repository will not work on any arbitrary OS version -- in practice, the only OS versions we expect them to work on are the exact ones that we use on our CI system. (And only to the extent covered by our tests.)

Set the minimum deployment target of stdlib binaries to match whatever OS versions we use on ci.swift.org. (This will need to be updated whenever we upgrade the CI machines.)

Note that while we don't check for a maximum deployment target, generated stdlib binaries aren't expected to be forward compatible with future OS releases, either. (Not even minor releases.) This is because OS frameworks are very likely to depend on new APIs introduced in their corresponding Swift stdlib release that would be missing from binaries built from an older codebase.

Also note that this deployment target does not affect our ability to build apps that need to deploy on earlier OSes -- it just affects the binary stdlib dylibs that we build. (The swiftinterface file isn't affected by the deployment target setting.)