[Serialization] Properly skip invalid destructors when allowing errors
7856f2d83da7fc213fb3c16676cd953b8e9ff369 only partially skipped writing
out destructors when they were invalid, ie. it skipped writing the decl
itself but not the records common to all decls. This would cause any
records on the destructor to be applied on the next serialized decl.

Make sure to skip serializing anything to do with the destructor when
it's invalid and does not have a class context.