Fix debug scope for SIL instructions created in LowerHopToActor
The pass uses SILBuilder instant to create instructions and does not
set SILDebugScope.
Due to this, SILVerifier correctly asserts that basic blocks contain non-contiguous
lexical scopes at -Onone.
The long term fix would be to replace SILBuilder instance in the pass with
SILBuilderContext and use SILBuilderWithScope so that such bugs do not come up.