[NFC] remove test use of `cmp` to compare swiftmodules
In the test infrastructure, I was using the binary
comparison tool `cmp` to check if the serialization
of a module can be deserialized and reserialized to
the exact same swiftmodule file. It turns out that
bitwise equality on the emitted files is not
guaranteed between those two methods of creating
a swiftmodule file is not always guaranteed.

Thus, this commit removes the uses of cmp to check
for this fragile guarantee.